requirements.txt
telethon~=1.36.0
aiohttp~=3.10.11
python-dotenv~=1.0.1
